requirements.txt
Django==2.1.3
django-enumfields==0.10.1
django-extensions==2.1.3
django-rest==0.0.1
django-rest-framework==0.1.0
django-shell-plus==1.1.7
djangorestframework==3.9.0
model-mommy==1.6.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
django-redis==4.10.0